minor fixes -2
