+++ package.json
@@ -12,6 +12,7 @@
         "chart.js": "^3.6.0",
         "copy-to-clipboard": "^3.3.1",
         "firebase": "^9.1.0",
+        "moment": "^2.29.1",
         "qrcode": "^1.4.4",
         "react": "^17.0.2",
         "react-chartjs-2": "^3.3.0",
