+++ src/component/ThemedButton.js
@@ -1,7 +1,7 @@
-const ThemedButton = ({ title, onClickHandler }) => {
+const ThemedButton = ({ title, onClickHandler, color, className }) => {
     return (
         <div
-            className="flex justify-between items-center text-xl py-2 px-3 rounded-xl bg-blue-800 text-white cursor-pointer"
+            className={`flex justify-between items-center text-xl py-2 px-3 rounded-xl ${color} text-white cursor-pointer ${className}`}
             onClick={onClickHandler}
         >
             <h1>{title}</h1>
