+++ src/configs/constants.js
@@ -1,4 +1,4 @@
-// export const BASE_URL = "https://shorty--urls-server.herokuapp.com/";
-export const BASE_URL = "http://localhost:4000/";
+export const BASE_URL = "https://shorty--urls-server.herokuapp.com/";
+// export const BASE_URL = "http://localhost:4000/";
 export const Authorization = "05d5f47ab1314523bffef0e918afd3cb";
 // export const BASE_URL = "http://5fe5-1-38-240-32.ngrok.io/";
