+++ src/screens/Home.js
@@ -229,25 +229,33 @@ export const Home = () => {
                                 color="bg-yellow-200"
                             />
                         </div>
-                        <h1 className="text-blue-500 text-2xl my-2 text-center border-t-2 pt-3">Most Clicked URLs</h1>
-                        <TopLinks
-                            title={meta?.top_three?.[0]?.title}
-                            url={meta?.top_three?.[0]?.url}
-                            short_url={meta?.top_three?.[0]?.short_url}
-                            color="#FFD700"
-                        />
-                        <TopLinks
-                            title={meta?.top_three?.[1]?.title}
-                            url={meta?.top_three?.[1]?.url}
-                            short_url={meta?.top_three?.[1]?.short_url}
-                            color="#C0C0C0"
-                        />
-                        <TopLinks
-                            title={meta?.top_three?.[2]?.title}
-                            url={meta?.top_three?.[2]?.url}
-                            short_url={meta?.top_three?.[2]?.short_url}
-                            color="#CD7F32"
-                        />
+                        {meta?.top_three?.length > 0 ? (
+                            <h1 className="text-blue-500 text-2xl my-2 text-center border-t-2 pt-3 m-2">Most Clicked URLs</h1>
+                        ) : null}
+                        {meta?.top_three?.[0] ? (
+                            <TopLinks
+                                title={meta?.top_three?.[0]?.title}
+                                url={meta?.top_three?.[0]?.url}
+                                short_url={meta?.top_three?.[0]?.short_url}
+                                color="#FFD700"
+                            />
+                        ) : null}
+                        {meta?.top_three?.[0] ? (
+                            <TopLinks
+                                title={meta?.top_three?.[1]?.title}
+                                url={meta?.top_three?.[1]?.url}
+                                short_url={meta?.top_three?.[1]?.short_url}
+                                color="#C0C0C0"
+                            />
+                        ) : null}
+                        {meta?.top_three?.[0] ? (
+                            <TopLinks
+                                title={meta?.top_three?.[2]?.title}
+                                url={meta?.top_three?.[2]?.url}
+                                short_url={meta?.top_three?.[2]?.short_url}
+                                color="#CD7F32"
+                            />
+                        ) : null}
                     </div>
                     <div className="flex flex-col border-2 p-3 rounded-xl" style={{ boxShadow: "0px 0px 15px 0.5px blue" }}>
                         <div className="flex items-center">
